api/pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "buildflow-api"
version = "0.1.0"
description = "BuildFlow AI backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "sqlalchemy>=2.0.41,<3.0.0",
  "alembic>=1.16.0,<2.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "openai>=1.0.0,<3.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0,<9.0.0",
  "httpx>=0.28.0,<1.0.0"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]